+++ src/core/types/flows/withdrawals.ts
@@ -1,13 +1,14 @@
 // src/types/flows/withdrawals.ts
-import type { Address, Hex, UInt } from '../primitives';
+
+import type { Address, Hex } from '../primitives';
 import type { ApprovalNeed, Plan, Handle } from './base';
 
 /** Input */
 export interface WithdrawParams {
   token: Address;
-  amount: UInt;
+  amount: bigint;
   to?: Address;
-  l2GasLimit?: UInt;
+  l2GasLimit?: bigint;
 }
 
 /** Routes */
@@ -17,7 +18,7 @@ export type WithdrawRoute = 'eth' | 'erc20';
 export interface WithdrawQuote {
   route: WithdrawRoute;
   approvalsNeeded: readonly ApprovalNeed[];
-  suggestedL2GasLimit: UInt;
+  suggestedL2GasLimit: bigint;
 }
 
 /** Plan (Tx generic) */
@@ -57,12 +58,13 @@ type WithdrawalPhase =
   | 'L2_PENDING' // tx not in an L2 block yet
   | 'L2_INCLUDED' // we have the L2 receipt
   | 'PENDING' // inclusion known; proof data not yet derivable/available
-  | 'READY_TO_FINALIZE' // we can construct proof params; nullifier not set
+  | 'READY_TO_FINALIZE' // Ready to call finalize on L1
   | 'FINALIZING' // L1 tx sent but not picked up yet
-  | 'FINALIZED' // nullifier says done
+  | 'FINALIZED' // L2-L1 tx finalized on L1
   | 'FINALIZE_FAILED' // prior L1 finalize reverted
   | 'UNKNOWN';
 
+// Withdrawal Status
 export type WithdrawalStatus = {
   phase: WithdrawalPhase;
   l2TxHash: Hex;
@@ -70,6 +72,8 @@ export type WithdrawalStatus = {
   key?: WithdrawalKey;
 };
 
+// Finalization readiness states
+// Used for `status()`
 export type FinalizeReadiness =
   | { kind: 'READY' }
   | { kind: 'FINALIZED' }
@@ -86,12 +90,14 @@ export type FinalizeReadiness =
       detail?: string;
     };
 
+// Parsed L1 receipt and logs
 export type ParsedLog = {
   address: string;
   topics: Hex[];
   data: Hex;
 };
 
+// Parsed L1 receipt
 export type ParsedReceipt = {
   logs: ParsedLog[];
 };
